Added device_duplicate_value services API option.
An integer value that specifies the status of a duplicate scan when scanning on-device and only checks against scans currently stored on the device. Set to 1 to make the status of duplicate scans actually on the device valid and set to 0 to make them invalid. The default status of duplicates on-device are valid (e.g. value set to 1). This only applies when the validation_method is set to record, ondevicerecord, database or ondevicedatabase. NOTE: if the duplicate_value option is set to 0 (invalid) then this option is overridden.